feat: restructure MLXServer into tools/server with full OpenAI compat
Move server from Sources/MLXServer/main.swift (1500 lines) to
tools/server/ with logical module split:

- MLXServerApp.swift — entry point, CLI args, model loading
- ServerTypes.swift — OpenAI request/response types, JSON codables
- ServerHTTP.swift — HTTP server, routing, keep-alive, SSE helpers
- ServerChat.swift — chat completions + text completions handlers
- ServerCache.swift — multi-session KV cache with LCP matching
- ServerSlots.swift — parallel inference slot manager with FIFO queue
- ServerTools.swift — tool call XML parsing (auto-detected per model)

Key fixes and features:
- Fix streaming hang: Connection: close for SSE, socket close after [DONE]
- Fix sequential requests: second request no longer blocks
- Fix tool call loop: pass tool role through to template natively
- Fix think leak: always-on thinking with proper stripping
- Fix tool call args: JSON-encoded string per OpenAI spec
- Add repetition penalty (default 1.1) to prevent looping
- Add all sampling params: top_k, min_p, frequency/presence/repeat penalty
- Add --reasoning on|off flag (like llama-server)
- Add --kv-bits, --ctx-size, --n-predict CLI flags
- Auto-detect tool call format from model_type (no hardcoding)
- Enable thinking for all requests, strip from content, expose as
  reasoning_content
